Add empty object to expressionStatements.
The idea is next: when you write
```
module.export = {
    42: 42
}
```

You assume that you are in node env, so you have global object module,
but we as you see in this code it doesn't exports anything because of typo.
`export` -> `exports`
so in node env you will get undefined
but as we want to be nice eval and eval expressions without module.exports:
```
({ 42: 42 })
```
we try to do it and get an Error, because code have no `module` to write `export` prop in it
So our desicion is to temporaly polute global object with `module`.